Cleanups again
But this time they work :D

+++ src/Network.h
@@ -2,11 +2,9 @@
 
 #include <algorithm>
 #include <array>
-#include <cstdint>
 #include <cstdio>
-#include <iostream>
 #include <ranges>
-#include <string>
+#include <string_view>
 
 #include <libopencm3/stm32/dma.h>
 #include <libopencm3/stm32/usart.h>
@@ -70,7 +68,7 @@ public:
             // Enable usart interrupts
             usart->error_interrupt(true);
             // Enable usart DMA
-            usart->enable_rx_dma(reinterpret_cast<uint32_t>(buf.data()), buf.size());
+            usart->enable_rx_dma(reinterpret_cast<uintptr_t>(buf.data()), buf.size());
 
             reset_pin->port->set_pins(reset_pin->pin_nro);
             bluepill::async_wait_ms(reset_time); // Wait a bit so the ESP8266 has time to reset
@@ -103,7 +101,7 @@ public:
             // Enable usart interrupts
             usart->error_interrupt(true);
             // Enable usart DMA
-            usart->enable_rx_dma(reinterpret_cast<uint32_t>(buf.data()), buf.size());
+            usart->enable_rx_dma(reinterpret_cast<uintptr_t>(buf.data()), buf.size());
 
             send_command_dont_care("AT+RST");
             bluepill::async_wait_ms(reset_time); // Wait a bit so the ESP8266 has time to reset
@@ -192,7 +190,7 @@ public:
     {
         if (ap_connected && socket_connected(id)) {
             char size_str[6];
-            sprintf(size_str, "%d", data.size());
+            sprintf(size_str, "%zu", data.size());
             send_raw("AT+CIPSEND=");
             send_command_dont_care(size_str);
             bluepill::async_wait_ms(5);
@@ -261,7 +259,7 @@ private:
         // Enable usart interrupts
         usart->error_interrupt(true);
         // Enable usart DMA
-        usart->enable_rx_dma(reinterpret_cast<uint32_t>(response_buf.data()), response_buf.size());
+        usart->enable_rx_dma(reinterpret_cast<uintptr_t>(response_buf.data()), response_buf.size());
 
         // Send the command
         send_command_dont_care(cmd);
